Generate an alternate package for tree-shaking to simplify usage
ESM imports of `chart.js` will no longer require registration of
controllers, scales, elements, or plugins. Instead, everything is
pre-registered.

To use tree shaking, imports will need to be made from the
'chart.js/tree' package and will require registration.
