+++ src/index.esm.js
@@ -1,3 +1,4 @@
+// Export all of these as we do in the tree-shakeable version
 export * from './controllers';
 export * from './core';
 export * from './elements';
@@ -5,8 +6,13 @@ export * from './platform';
 export * from './plugins';
 export * from './scales';
 
-// Exports with names so that users can register faster
-export * as controllers from './controllers';
-export * as elements from './elements';
-export * as plugins from './plugins';
-export * as scales from './scales';
+import {Chart} from './core';
+import * as controllers from './controllers';
+import * as elements from './elements';
+import * as plugins from './plugins';
+import * as scales from './scales';
+
+Chart.register(controllers, elements, plugins, scales);
+
+// Since everything is already registered, provide a default Chart export for convenience
+export default Chart;
