+++ src/index.shakeable.esm.js
@@ -1,0 +1,6 @@
+export * from './controllers';
+export * from './core';
+export * from './elements';
+export * from './platform';
+export * from './plugins';
+export * from './scales';
